requirements.txt
idna==3.11
playwright==1.58.0
pyee==13.0.1
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3